Fix equations of Het1D Steady State and Transient
