+++ examples/het1d_transient.py
@@ -7,7 +7,7 @@ if __name__ == "__main__":
     h = TransientHeterogeneous1DReactor(os.path.join('files', 'H2-O2-Rh.yaml'), 'gas', 'Rh_surface')
     h.set_length([0, 0.5, 1.0, 1.5, 2.0], 'm')
     h.set_pressure(5, 'bar')
-    h.set_catalytic_load(10, '1/m')
+    h.set_catalytic_load(35, '1/m')
     h.set_volumetric_flow_rate(15., 'm3/h')
     h.set_inlet_temperature(250, 'degC')
     h.set_inlet_mass_fraction({'O2': 0.4, 'AR': 0.5, 'H2': 0.1})
@@ -17,7 +17,7 @@ if __name__ == "__main__":
     h.set_solid_thermal_conductivity(2.5, 'W/m/degK')
     h.set_initial_solid_temperature(250, 'degC')
     h.set_energy(True)
-    h.set_gas_diffusion(False)
+    h.set_gas_diffusion(True)
     h.set_verbosity(False)
     h.set_initial_mass_fraction({'O2': 0.4, 'AR': 0.5, 'H2': 0.1})
     h.set_initial_temperature(250, 'degC')
